feat: #WB-1953, align color picker and its CSS rules
